Rename scatteplot.js to scatterplot.js
